+++ framework/core/src/mappers/index.ts
@@ -1,5 +1,2 @@
 export * from './models/requestDtoMapper.model';
 export * from './models/responseDtoMapper.model';
-export * from './serviceSchemaResolver';
-export * from './types/mappers.types';
-export * from './types/serviceSchemaResolver.types';