+++ framework/infrastructure/S3/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @forklaunch/infrastructure-s3
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/core@0.9.20
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/core@0.9.19
+
 ## 0.0.26
 
 ### Patch Changes